+++ tests/Unit/AbstractClientTest.php
@@ -25,7 +25,7 @@ use CrowdSec\CapiClient\Watcher;
 
 /**
  * @uses \CrowdSec\CapiClient\HttpMessage\Response
- * @uses \CrowdSec\CapiClient\Configuration::getConfigTreeBuilder
+ * @uses \CrowdSec\CapiClient\Configuration\Watcher::getConfigTreeBuilder
  * @uses \CrowdSec\CapiClient\Watcher::formatUserAgent
  * @uses \CrowdSec\CapiClient\Storage\FileStorage::__construct
  * @uses \CrowdSec\CapiClient\RequestHandler\AbstractRequestHandler::__construct